Colocar las entidades en un mismo archivo y arreglar algunos errores
