--- backend/app/entities/group_entity.py
@@ -1,4 +1,0 @@
-class GroupEntity:
-    def __init__(self, name, id=None):
-        self.id = id
-        self.name = name
