--- backend/app/entities/local_entity.py
@@ -1,5 +1,0 @@
-class LocalEntity:
-    def __init__(self, name, size, id=None):
-        self.id = id
-        self.name = name
-        self.size = size
