--- backend/app/entities/notification_entity.py
@@ -1,5 +1,0 @@
-class NotificationEntity:
-    def __init__(self, date, text, id=None):
-        self.id = id
-        self.date = date
-        self.text = text
