--- backend/app/entities/option_entity.py
@@ -1,5 +1,0 @@
-class OptionEntity:
-    def __init__(self, text, votation_id, id=None):
-        self.id = id
-        self.text = text
-        self.votation_id = votation_id
