--- backend/app/entities/permission_entity.py
@@ -1,4 +1,0 @@
-class PermissionEntity:
-    def __init__(self, code, id=None):
-        self.id = id
-        self.code = code
