--- backend/app/entities/resource_entity.py
@@ -1,4 +1,0 @@
-class ResourceEntity:
-    def __init__(self, name, id=None):
-        self.id = id
-        self.name = name
