--- backend/app/entities/role_entity.py
@@ -1,5 +1,0 @@
-class RoleEntity:
-    def __init__(self, name, id=None):
-        self.id = id
-        self.name = name
-
