--- backend/app/entities/tag_entity.py
@@ -1,4 +1,0 @@
-class TagEntity:
-    def __init__(self, text, id=None):
-        self.id = id
-        self.text = text
