--- backend/app/entities/type_entity.py
@@ -1,4 +1,0 @@
-class TypeEntity:
-    def __init__(self, name, id=None):
-        self.id = id
-        self.name = name
