--- backend/app/entities/votation_entity.py
@@ -1,5 +1,0 @@
-class VotationEntity:
-    def __init__(self, title, text, id=None):
-        self.id = id
-        self.title = title
-        self.text = text
